Project manager
closes #45
closes #23
closes #26

+++ Source/Engine/Core/src/Input/InputManager.cpp
@@ -58,7 +58,7 @@ namespace Core
 				result = getDeviceForId(event.deviceId);
 				break;
 			default:
-				Logger::warn("InputManager : Unknown device '%zu'", event.deviceType);
+				Logger::warn("InputManager : Unknown device '%zu'", (unsigned int) event.deviceType);
 				break;
 		}
